# docs/submission/openapi_service_control.yaml
openapi: 3.0.3
info:
  title: AI Service Control API
  version: "0.5.0"
  description: >
    Go Echo contract for the KHU AI application automation Agent. The core flow
    joins ApplicationProfile and ResourceRecommendation messages, authorizes the
    built-in AIApplicationAutomationAgent through Agent Registry capability and
    bounded Action policy, decides DEPLOY, REJECT, or RETRY, applies deterministic
    Go Guard checks, and emits a platform-neutral Desired Deployment Spec inside
    deployment.create.request. Deployment execution and platform-specific Manifest
    conversion remain external responsibilities. Legacy ControlRun, runtime Agent
    execution, and post-deployment controls are retained as optional PoC tools.
servers:
  - url: http://127.0.0.1:18080
paths:
...
  /api/v1/agent-control/application-analysis-requests:
    post:
      operationId: PostApplicationAnalysisRequest
      summary: Start the Agent flow from a Common JSON analysis request
      description: >
        Accepts application.analysis.request, preserves correlation_id and
        trace_id, and executes Requirement Analyzer, Resource Recommender,
        Agent Registry authorization, deployment decision, and Go Guard.
        message_id is the idempotency key. An identical replay returns the
        original run without generating another deployment.create.request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationAnalysisRequestEnvelope"
      responses:
        "201":
          description: Analysis request processed for the first time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRun"
        "200":
          description: Existing AutomationRun returned for an idempotent replay
          headers:
            Idempotent-Replayed:
              schema: {type: string, enum: ["true"]}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRun"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: The message_id already exists with a different payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/agent-control/automation-runs/{run_id}:
    get:
      operationId: GetAgentControlAutomationRun
      summary: Get one automatic three-stage Agent run
      parameters:
        - name: run_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Requirement analysis, resource recommendation, Agent decision, Guard evidence, and Desired Deployment Spec
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRun"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/agent-control/application-contexts:
    post:
      summary: Receive an Application Context message
      description: Stores an application.context.created Common JSON v1.0 message and waits for its matching Resource Recommendation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationContextEnvelope"
      responses:
        "202":
          description: Application Context accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentControlFlow"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/agent-control/resource-recommendations:
    post:
      summary: Receive a Resource Recommendation message
      description: Stores a resource.recommendation.created Common JSON v1.0 message, joins it by correlation_id, trace_id, and profile_id, verifies the automation Agent against Agent Registry, and then performs decision and Guard processing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceRecommendationEnvelope"
      responses:
        "202":
          description: Resource Recommendation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentControlFlow"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/agent-control/deployment-status:
    post:
      summary: Receive a deployment status message
      description: Attaches a deployment.status.changed Common JSON v1.0 message to its correlation flow and summarizes the current outcome.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentStatusEnvelope"
      responses:
        "202":
          description: Deployment status accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentControlFlow"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/agent-control/optimization-feedback:
    post:
      summary: Receive optimization feedback
      description: >
        Attaches an optimization.feedback.created Common JSON v1.0 message,
        joins it with deployment.status.changed, and records the selected
        Operation Optimization Agent evidence. A resulting scaling decision is
        a guarded recommendation only; geon does not execute VM, Kubernetes,
        AppDeploy, or other infrastructure control from this endpoint.
      parameters:
        - name: operation_agent
          in: query
          required: false
          description: >
            Optional Registry-authorized Operation Optimization Agent name.
            When omitted, the Registry default is used.
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptimizationFeedbackEnvelope"
      responses:
        "202":
          description: Optimization feedback accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentControlFlow"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/agent-control/flows/{correlation_id}:
    get:
      summary: Get one Agent Control input flow
      parameters:
        - name: correlation_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Correlated input flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentControlFlow"
        "404":
          description: Flow not found
    delete:
      summary: Delete one Agent Control experiment flow
      description: Deletes one in-memory experiment record by correlation_id.
      parameters:
        - name: correlation_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Experiment flow deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted, correlation_id]
                properties:
                  deleted: {type: integer, enum: [1]}
                  correlation_id: {type: string}
        "404":
          description: Flow not found
  /api/v1/agent-control/flows/{correlation_id}/reasoning-comparisons:
    post:
      summary: Compare simple and validated Qwen reasoning
      description: Returns the deterministic rule result, raw Qwen proposal, and the final proposal after Go Guard validation.
      parameters:
        - name: correlation_id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentControlReasoningComparisonRequest"
      responses:
        "200":
          description: Reasoning comparison, including provider_unavailable when the configured LLM cannot be reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningComparison"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    AutonomyAdminToken:
      type: http
      scheme: bearer
      bearerFormat: opaque admin token
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested record was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    AutomationRunInput:
      type: object
      required: [input_type]
      properties:
        input_type:
          type: string
          enum: [natural_language, structured]
        request:
          type: string
          description: Natural-language deployment and operation request
        requested_by: {type: string}
        decision_agent:
          type: string
          description: Optional Agent Registry name. When omitted, the ai_application_automation default is used.
        app_spec:
          $ref: "#/components/schemas/StructuredAppSpec"
      description: natural_language requires request; structured requires app_spec.
...
    AutomationRun:
      type: object
      required: [run_id, correlation_id, trace_id, status, input, created_at, updated_at]
      properties:
        run_id: {type: string}
        correlation_id: {type: string}
        trace_id: {type: string}
        status: {type: string, enum: [COMPLETED, FAILED]}
        input:
          $ref: "#/components/schemas/AutomationRunInput"
        requirement_analysis:
          $ref: "#/components/schemas/RequirementAnalysisResult"
        resource_recommendation:
          $ref: "#/components/schemas/RecommendationResult"
        flow:
          $ref: "#/components/schemas/AgentControlFlow"
        desired_deployment_spec:
          $ref: "#/components/schemas/DesiredDeploymentSpec"
        deployment_submission:
          $ref: "#/components/schemas/DeploymentSubmission"
        error_code: {type: string}
        error_message: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    DeploymentSubmission:
      type: object
      required: [adapter, status, simulated, request_id, submitted_at]
      properties:
        adapter:
          type: string
          enum: [mock, handoff]
        status:
          type: string
          enum: [SIMULATED, READY, FAILED]
        simulated:
          type: boolean
        request_id:
          type: string
        submitted_at:
          type: string
          format: date-time
        error_code:
          type: string
        error_message:
          type: string
    RequirementAnalysisResult:
      type: object
      required: [application_profile, model_recommendation, mode, evidence]
      properties:
        application_profile:
          $ref: "#/components/schemas/ApplicationProfile"
        model_recommendation:
          $ref: "#/components/schemas/ApplicationModelRecommendation"
        mode: {type: string, enum: [local_rule, structured, qwen]}
        evidence:
          $ref: "#/components/schemas/RequirementAnalysisEvidence"
...
    RecommendationResult:
      type: object
      required: [resource_recommendation, evidence]
      properties:
        resource_recommendation:
          $ref: "#/components/schemas/ResourceRecommendation"
        evidence:
          $ref: "#/components/schemas/RecommendationEvidence"
...
    DesiredDeploymentSpec:
      type: object
      required:
        - spec_version
        - decision_id
        - application
        - target_runtime
        - desired_infrastructure
        - inference_configuration
        - metadata
      properties:
        spec_version: {type: string, enum: ["1.0"]}
        decision_id: {type: string}
        application:
          type: object
          required: [app_id, app_version]
          properties:
            app_id: {type: string}
            app_version: {type: string}
        target_runtime: {type: string, enum: [VM]}
        desired_infrastructure:
          $ref: "#/components/schemas/DesiredInfrastructure"
        inference_configuration:
          type: object
          additionalProperties: true
        runtime:
          type: object
          additionalProperties: true
        policy_hints: {type: array, items: {type: string}}
        metadata:
          type: object
          required: [profile_id]
          properties:
            profile_id: {type: string}
...
    CommonMessageEnvelope:
      type: object
      required:
        - contract_version
        - message_id
        - message_type
        - occurred_at
        - correlation_id
        - trace_id
        - source
        - target
      properties:
        contract_version: {type: string, enum: ["1.0"]}
        message_id: {type: string}
        message_type: {type: string}
        occurred_at: {type: string, format: date-time}
        correlation_id: {type: string}
        trace_id: {type: string}
        causation_id: {type: string}
        source:
          $ref: "#/components/schemas/CommonMessageEndpoint"
        target:
          $ref: "#/components/schemas/CommonMessageEndpoint"
    ApplicationArtifact:
      type: object
      required: [type, uri]
      properties:
        type: {type: string, example: OCI_IMAGE}
        uri:
          type: string
          example: docker://registry.example.org/chat-service:1.0.0
        entrypoint:
          type: array
          items: {type: string}
    ApplicationAnalysisRequestEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [application.analysis.request]
            data:
              type: object
              required: [application]
              properties:
                application:
                  type: object
                  required:
                    - app_id
                    - app_version
                    - artifact
                    - user_request
                  properties:
                    app_id: {type: string, example: chat-service}
                    app_version: {type: string, example: 1.0.0}
                    artifact:
                      $ref: "#/components/schemas/ApplicationArtifact"
                    user_request:
                      type: string
                      example: 동시 사용자 20명, p95 지연시간 2초 이하로 배포
                    declared_spec:
                      type: object
                      properties:
                        expected_rps: {type: number, minimum: 0, example: 5}
                        max_input_tokens: {type: integer, minimum: 0, example: 4096}
                    labels:
                      type: object
                      additionalProperties: {type: string}
    ApplicationContextEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [application.context.created]
            data:
              type: object
              required: [application_profile, model_recommendation]
              properties:
                application_profile:
                  $ref: "#/components/schemas/ApplicationProfile"
                model_recommendation:
                  $ref: "#/components/schemas/ApplicationModelRecommendation"
    ApplicationProfile:
      type: object
      required: [profile_id, app_id, app_version, requirements]
      properties:
        profile_id: {type: string}
        app_id: {type: string}
        app_version: {type: string}
        artifact:
          type: object
          properties:
            type: {type: string}
            uri: {type: string}
            entrypoint:
              type: array
              items: {type: string}
        workload:
          type: object
          properties:
            task_type: {type: string}
            request_pattern: {type: string}
            expected_rps: {type: number, minimum: 0}
        requirements:
          $ref: "#/components/schemas/ApplicationRequirements"
        analysis:
          type: object
          properties:
            confidence: {type: number, minimum: 0, maximum: 1}
            assumptions: {type: array, items: {type: string}}
            missing_fields: {type: array, items: {type: string}}
            warnings: {type: array, items: {type: string}}
...
    ApplicationModelRecommendation:
      type: object
      required: [selected_model, inference_configuration]
      properties:
        recommendation_id: {type: string}
        selected_model:
          type: object
          required: [model_id, model_version, source]
          properties:
            model_id: {type: string}
            model_version: {type: string}
            source: {type: string}
        inference_configuration:
          type: object
          required:
            - runtime_engine
            - precision
            - max_batch_size
            - max_concurrency
            - tensor_parallel_size
            - replicas
          properties:
            runtime_engine: {type: string}
            precision: {type: string}
            max_batch_size: {type: integer, minimum: 1}
            max_concurrency: {type: integer, minimum: 1}
            tensor_parallel_size: {type: integer, minimum: 1}
            replicas: {type: integer, minimum: 1}
    ResourceRecommendationEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [resource.recommendation.created]
            data:
              type: object
              required: [resource_recommendation]
              properties:
                resource_recommendation:
                  $ref: "#/components/schemas/ResourceRecommendation"
...
    DeploymentStatusEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [deployment.status.changed]
            data:
              type: object
              required: [deployment_status]
              properties:
                deployment_status:
                  $ref: "#/components/schemas/DeploymentStatus"
...
    OptimizationFeedbackEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [optimization.feedback.created]
            data:
              type: object
              required: [optimization_feedback]
              properties:
                optimization_feedback:
                  $ref: "#/components/schemas/OptimizationFeedback"
...
    AgentControlFlow:
      type: object
      required: [correlation_id, trace_id, state, updated_at]
      properties:
        correlation_id: {type: string}
        trace_id: {type: string}
        automation_run_id: {type: string}
        requested_decision_agent: {type: string}
        requested_operation_agent:
          type: string
          description: Operation Optimization Agent requested for the joined deployment status and Feedback.
        profile_id: {type: string}
        state:
          type: string
          enum:
            - WAITING_FOR_APPLICATION_CONTEXT
            - WAITING_FOR_RESOURCE_RECOMMENDATION
            - READY
            - AGENT_AUTHORIZATION_REJECTED
            - AGENT_EXECUTION_FAILED
            - AGENT_RESULT_REJECTED
            - DEPLOY_APPROVED
            - REJECTED
            - RETRY_REQUIRED
        application_context:
          $ref: "#/components/schemas/ApplicationContextEnvelope"
        resource_recommendation:
          $ref: "#/components/schemas/ResourceRecommendationEnvelope"
        agent_authorization:
          $ref: "#/components/schemas/AgentAuthorization"
        agent_execution:
          $ref: "#/components/schemas/DecisionAgentResult"
        decision:
          $ref: "#/components/schemas/AutomationDecision"
        deployment_plan:
          $ref: "#/components/schemas/AgentControlDeploymentPlan"
        guard:
          $ref: "#/components/schemas/AgentControlGuardResult"
        deployment_request:
          $ref: "#/components/schemas/DeploymentCreateRequestEnvelope"
        deployment_status:
          $ref: "#/components/schemas/DeploymentStatusEnvelope"
        optimization_feedback:
          $ref: "#/components/schemas/OptimizationFeedbackEnvelope"
        feedback_summary:
          $ref: "#/components/schemas/FeedbackSummary"
        operation_agent_execution:
          $ref: "#/components/schemas/OperationOptimizationResult"
        scaling_decision:
          $ref: "#/components/schemas/ScalingDecision"
        reasoning_comparison:
          $ref: "#/components/schemas/ReasoningComparison"
        updated_at: {type: string, format: date-time}
    AgentAuthorization:
      type: object
      required: [agent_name, capability, action, authorized, reason]
      properties:
        agent_name:
          type: string
          description: Selected configuration or Runtime Agent name
        capability:
          type: string
          enum: [ai_application_automation]
        action:
          type: string
          enum: [generate_deployment_decision]
        authorized: {type: boolean}
        reason: {type: string}
...
    AutomationDecision:
      type: object
      required:
        - decision_id
        - action
        - reason
        - reasoning_mode
        - confidence
        - created_at
      properties:
        decision_id: {type: string}
        action: {type: string, enum: [DEPLOY, REJECT, RETRY]}
        reason: {type: string}
        reasoning_mode: {type: string, enum: [rule_based]}
        confidence: {type: number, minimum: 0, maximum: 1}
        selected_candidate_id: {type: string}
        correction_request:
          $ref: "#/components/schemas/CorrectionRequest"
        created_at: {type: string, format: date-time}
...
    AgentControlReasoningComparisonRequest:
      type: object
      required: [candidate_id]
      properties:
        candidate_id: {type: string, example: qwen3.5-ops-planner}
    ReasoningComparison:
      type: object
      required:
        - correlation_id
        - candidate_id
        - rule_based
        - simple_inference
        - validated_inference
        - agreement
        - created_at
      properties:
        correlation_id: {type: string}
        candidate_id: {type: string}
        rule_based:
          $ref: "#/components/schemas/ReasoningTrial"
        simple_inference:
          $ref: "#/components/schemas/ReasoningTrial"
        validated_inference:
          $ref: "#/components/schemas/ReasoningTrial"
        agreement:
          type: object
          required: [action_match, candidate_match]
          properties:
            action_match: {type: boolean}
            candidate_match: {type: boolean}
        created_at: {type: string, format: date-time}
...
    ErrorResponse:
      type: object
      required: [valid, message]
      properties:
        valid: {type: boolean, example: false}
        message: {type: string}
...